Result의 확장 함수 onSuccess, onFailure 추가
